refactor: added more docs

+++ dev/generate_all_feature_definitions.py
@@ -7,16 +7,22 @@ import shutil
 
 
 def generate_all_Feature_definitions(
-    feature_definitions_dir: str, output_dir: str
+    feature_definitions_dir: str, output_dir: str ,remove_old_content: bool = True,
 ) -> None:
-    
 
     for feature_name in os.listdir(feature_definitions_dir):
-        if os.path.exists(os.path.join(
+        if remove_old_content and os.path.exists(os.path.join(
                     "test", feature_name
                 ) ):
             shutil.rmtree(os.path.join(
                         "test", feature_name
+                    )
+                    )
+        if remove_old_content and os.path.exists(os.path.join(
+                    "src", feature_name
+                ) ):
+            shutil.rmtree(os.path.join(
+                        "src", feature_name
                     )
                     )
